moved WAXtax to top directory to simplify usage
