+++ pyproject.toml
@@ -16,6 +16,3 @@ black = "^21.9b0"
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
-
-[tool.poetry.scripts]
-waxtax = "waxtax.run_with_poetry:main"
