--- src/waxtax/__init__.py
@@ -1,4 +1,0 @@
-"""WAXTax Transaction Exporter.
-
-Author: SixPM Software
-"""
