+++ waxtax/__init__.py
@@ -1,0 +1,4 @@
+"""WAXTax Transaction Exporter.
+
+Author: SixPM Software
+"""
