+++ waxtax/run_with_poetry.py
@@ -1,0 +1,8 @@
+import subprocess
+import sys
+
+
+def main():
+    subprocess.run(
+        ["poetry", "run", "python", "src/waxtax"], stderr=sys.stderr, stdout=sys.stdout
+    )
